begin norminette

+++ Includes/fdf.h
@@ -6,7 +6,7 @@
 /*   By: jvigny <jvigny@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/19 15:32:17 by jvigny            #+#    #+#             */
-/*   Updated: 2023/02/07 16:27:21 by jvigny           ###   ########.fr       */
+/*   Updated: 2023/02/08 11:25:25 by jvigny           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -97,15 +97,15 @@ typedef struct s_coordonnee_3d
 typedef struct s_move
 {
 	short	zoom_in;
-	int	zoom_out;
-	int	translate_up;
-	int	translate_down;
-	int	translate_left;
-	int	translate_right;
-	int	rotate_a;
-	int	rotate_w;
-	int	rotate_d;
-	int	rotate_s;
+	short	zoom_out;
+	short	translate_up;
+	short	translate_down;
+	short	translate_left;
+	short	translate_right;
+	short	rotate_a;
+	short	rotate_w;
+	short	rotate_d;
+	short	rotate_s;
 }	t_move;
 
 typedef struct s_data
@@ -115,7 +115,6 @@ typedef struct s_data
 	int		octets_per_pixel;
 	int		line_length;
 	int		endian;
-	//reste 4 octets sur la page
 }	t_data;
 
 typedef struct s_game
@@ -155,21 +154,27 @@ void			fill_color(t_game *game);
 
 // ------ Draw ------
 void			my_mlx_pixel_put(t_data *data, int x, int y, int color);
-void			draw_line(t_coordonnee_3d pt1, t_coordonnee_3d pt2, t_data *data);
+void			draw_line(t_coordonnee_3d pt1, t_coordonnee_3d pt2,
+					t_data *data);
 void			link_point(t_game *game);
 t_color_f		gradient(t_color color1, t_color color2, float y1, float y2);
 
 // ------ Rotate ------
-void			rotate_plan_x(t_coordonnee_3d *tab_3d, unsigned int len, double rotation);
-void			rotate_plan_y(t_coordonnee_3d *tab_3d, unsigned int len, double rotation);
-void			rotate_plan_z(t_coordonnee_3d *tab_3d, unsigned int len, double rotation);
-void			rotate_plan_isometrique(t_coordonnee_3d *tab_3d, t_coordonnee_3d *tab_const, unsigned int len);
-void			rotate_plan_cabinet(t_coordonnee_3d *tab_3d, t_coordonnee_3d *tab_const, unsigned int len);
+void			rotate_plan_x(t_coordonnee_3d *tab_3d, unsigned int len,
+					double rotation);
+void			rotate_plan_y(t_coordonnee_3d *tab_3d, unsigned int len,
+					double rotation);
+void			rotate_plan_z(t_coordonnee_3d *tab_3d, unsigned int len,
+					double rotation);
+void			rotate_plan_isometrique(t_coordonnee_3d *tab_3d,
+					t_coordonnee_3d *tab_const, unsigned int len);
+void			rotate_plan_cabinet(t_coordonnee_3d *tab_3d,
+					t_coordonnee_3d *tab_const, unsigned int len);
 
 // ----- Keycode -----
 int				key(int keycode, t_game *mlx);
 int				key_release(int keycode, t_game *game);
-int				mouse(int keycode,int x, int y, t_game *game);
+int				mouse(int keycode, int x, int y, t_game *game);
 int				ft_translate(t_game *game, t_coordonnee_3d origine);
 int				ft_move(t_game *mlx);
 int				ft_close(t_game *mlx);
@@ -178,14 +183,15 @@ int				ft_rotate(t_game *game, float angle, char plan);
 int				ft_projection_iso(t_game *game);
 int				ft_projection_cabinet(t_game *game);
 
-
 // ------ Parsing -----
 unsigned int	ft_fdflen(char *str, unsigned int *len_line, t_game *game);
-void			ft_fill_tab(char *str, t_coordonnee_3d	*tab, int len_line, t_game *game);
+void			ft_fill_tab(char *str, t_coordonnee_3d	*tab, int len_line,
+					t_game *game);
 int				parsing(int argc, char **argv, t_game *game);
 
 // ------ Zoom ------
-float			center_plan(t_coordonnee_3d *tab, unsigned int len, t_coordonnee_3d *origine);
+float			center_plan(t_coordonnee_3d *tab, unsigned int len,
+					t_coordonnee_3d *origine);
 void			apply_zoom(t_game *game, float zoom);
 
 // ------ Translation ------
@@ -197,6 +203,7 @@ void			add_z(t_coordonnee_3d *tab, unsigned int len);
 
 // ------ Affichage -------
 void			ft_print(t_coordonnee_3d *tab, unsigned int len);
-void			ft_printf_point_of_tab(t_coordonnee_3d *tab_3d, unsigned int len, t_data *data);
+void			ft_printf_point_of_tab(t_coordonnee_3d *tab_3d,
+					unsigned int len, t_data *data);
 
 #endif
